+++ client/src/pages/Login/Login.js
@@ -1,42 +1,14 @@
 import React from "react";
 import { Redirect, useHistory } from "react-router-dom";
 import { connect } from "react-redux";
-import { makeStyles, createStyles } from "@material-ui/core/styles";
-import {
-  Grid,
-  Box,
-  Typography,
-  Hidden,
-  Button
-} from "@material-ui/core";
+import { Typography, Button } from "@material-ui/core";
 
 import LoginForm from "./LoginForm";
-import SideBanner from "../../components/SideBanner";
 import ConditionalElement from "../../components/ConditionalElement";
+import LoginTemplate from "../../components/Templates/LoginTemplate";
 import { login } from "../../store/utils/thunkCreators";
 
 import BubbleIcon from "../../components/BubbleIcon";
-
-const useStyles = makeStyles((theme) => createStyles({
-  root: {
-    display: "flex",
-    flex: 1,
-    flexDirection: "column",
-    alignItems: "flex-start",
-    justifyContent: "flex-start",
-    height: "100%"
-  },
-  formContainer: {
-    display: "flex",
-    flex: 1,
-    width: "60%",
-    margin: "0 auto",
-    marginTop: "10rem",
-    [theme.breakpoints.down('sm')]: {
-      marginTop: "2rem",
-    },
-  }
-}));
 
 const generateButtons = (history) => ([
   {
@@ -68,7 +40,6 @@ const generateButtons = (history) => ([
 ]);
 
 const Login = (props) => {
-  const classes = useStyles();
   const history = useHistory();
 
   const signupButtons = generateButtons(history);
@@ -79,42 +50,14 @@ const Login = (props) => {
     return <Redirect to="/home" />;
   }
 
-  const form = (
-    <Box className={classes.root}>
-      <Grid container justify="flex-end" alignItems="center">
-        <Hidden smDown>
-          <Box m={2}>
-            <Typography>Don't have an account?</Typography>
-          </Box>
-        </Hidden>
-        <Box m={2}>
-          <ConditionalElement elements={signupButtons} />
-        </Box>
-      </Grid>
-      <Box className={classes.formContainer}>
-        <Grid container>
-          <Hidden smUp>
-            <Box mx="auto" mb={2}>
-              <BubbleIcon fill="#3A8DFF" />
-            </Box>
-          </Hidden>
-          <LoginForm login={login} />
-        </Grid>
-      </Box>
-    </Box>
-  );
-
   return (
-    <>
-      <Hidden smDown>
-        <SideBanner >
-          {form}
-        </SideBanner>
-      </Hidden>
-      <Hidden smUp>
-        {form}
-      </Hidden>
-    </>
+    <LoginTemplate
+      formIcon={<BubbleIcon fill="#3A8DFF" />}
+      headerText={<Typography>Don't have an account?</Typography>}
+      headerButton={<ConditionalElement elements={signupButtons} />}
+    >
+      <LoginForm login={login} />
+    </LoginTemplate>
   );
 };
