+++ client/src/pages/Signup/Signup.js
@@ -1,40 +1,12 @@
 import React from "react";
 import { Redirect, useHistory } from "react-router-dom";
 import { connect } from "react-redux";
-import {
-  Grid,
-  Typography,
-  Box,
-  Hidden,
-  Button
-} from "@material-ui/core";
-import { makeStyles, createStyles } from "@material-ui/core/styles";
+import { Typography, Button } from "@material-ui/core";
 
 import SignupForm from "./SignupForm";
-import SideBanner from "../../components/SideBanner";
+import LoginTemplate from "../../components/Templates/LoginTemplate";
 import ConditionalElement from "../../components/ConditionalElement";
 import { register } from "../../store/utils/thunkCreators";
-
-const useStyles = makeStyles((theme) => createStyles({
-  root: {
-    display: "flex",
-    flex: 1,
-    flexDirection: "column",
-    alignItems: "flex-start",
-    justifyContent: "flex-start",
-    height: "100%"
-  },
-  formContainer: {
-    display: "flex",
-    flex: 1,
-    width: "60%",
-    margin: "0 auto",
-    marginTop: "10rem",
-    [theme.breakpoints.down('sm')]: {
-      marginTop: "2rem",
-    },
-  }
-}));
 
 const generateButtons = (history) => ([
   {
@@ -68,44 +40,19 @@ const generateButtons = (history) => ([
 const Signup = (props) => {
   const { user, register } = props;
   const history = useHistory();
-  const classes = useStyles();
-  const signupButtons = generateButtons(history);
+  const loginButtons = generateButtons(history);
 
   if (user.id) {
     return <Redirect to="/home" />;
   }
 
-  const form = (
-    <Box className={classes.root}>
-      <Grid container justify="flex-end" alignItems="center">
-        <Hidden smDown>
-          <Box m={2}>
-            <Typography>Already have an account?</Typography>
-          </Box>
-        </Hidden>
-        <Box m={2}>
-          <ConditionalElement elements={signupButtons} />
-        </Box>
-      </Grid>
-      <Box className={classes.formContainer}>
-        <Grid container>
-          <SignupForm register={register} />
-        </Grid>
-      </Box>
-    </Box>
-  );
-
   return (
-    <>
-      <Hidden smDown>
-        <SideBanner >
-          {form}
-        </SideBanner>
-      </Hidden>
-      <Hidden smUp>
-        {form}
-      </Hidden>
-    </>
+    <LoginTemplate
+      headerText={<Typography>Already have an account?</Typography>}
+      headerButton={<ConditionalElement elements={loginButtons} />}
+    >
+      <SignupForm register={register} />
+    </LoginTemplate>
   );
 };
